==> 04_Test_Locally.ipynb <==
# Test the deployed web application

In this notebook, we test the local web app running inside the Docker container we made previously.

import os
import numpy as np
import requests
import pandas as pd
import json
from timer import elapsed

from utilities import text_to_json
%load_ext dotenv
%dotenv

image_name = os.getenv('docker_login') + os.getenv('image_repo')
image_name

Run the Docker conatainer in the background and open port 80.

%%bash --bg -s "$image_name"
docker run -p 80:80 $1

Wait a few seconds for the application to spin up and then check that everything works.

import time
print('waiting for docker')
time.sleep(120)

!curl "http://0.0.0.0:80/"

!curl "http://0.0.0.0:80/version"

Now, let's use one of the duplicate questions to test our driver.

dupes_test_path = 'dupes_test.tsv'
dupes_test = pd.read_csv(dupes_test_path, sep='\t', encoding='latin1')
text_to_score = dupes_test.iloc[0,4]
text_to_score

jsontext = text_to_json(text_to_score)
jsontext[:100]

headers = {'content-type': 'application/json'}
%time r = requests.post('http://0.0.0.0:80/score', data=jsontext, headers=headers)
print(r)
r.json()

Let's try a few more duplicate questions and display their top 3 original matches.

dupes_to_score = dupes_test.iloc[:5,4]

url = 'http://0.0.0.0:80/score'
results = [requests.post(url, data=text_to_json(text), headers=headers) for text in dupes_to_score]

Let's print top 3 matches for each duplicate question.

[results[i].json()['result'][0][0:3] for i in range(0, len(results))]

Next let's quickly check what the request response performance is for the locally running Docker container.

text_data = list(map(text_to_json, dupes_to_score)) # Retrieve the text data

timer_results = [elapsed(requests.post, return_elapsed=True)(url, data=data, headers=headers)
                 for data in text_data]

timer_results

print('Average time taken: {0:4.2f} ms'.format(10**3 * np.mean(timer_results)))

%%bash
docker stop $(docker ps -q)

We can now [deploy our web application](05_Deploy_On_AKS.ipynb) on AKS.
