==> 09_Tear_Down.ipynb <==
# Tear it all down
Use this notebook to tear down the app and the cluster created by the tutorial.

Load the parameters defined in notebook 05_Deploy_On_AKS.

%load_ext dotenv
%dotenv

Once you are done with your cluster you can use the following two commands to destroy it all. First, delete the app.

!kubectl delete -f az-ml.json

Then delete the resource group. This also deletes the AKS cluster. This step may take multiple minutes.

!az group delete --name $resource_group -y
